fix it not working ig

+++ commands/image-editing/speed.js
@@ -2,7 +2,7 @@ import ImageCommand from "../../classes/imageCommand.js";
 
 class SpeedCommand extends ImageCommand {
   params() {
-    const speed = parseInt(this.options.multiplier ?? this.args[0]);
+    const speed = parseFloat(this.options.multiplier ?? this.args[0]);
     return {
       speed: isNaN(speed) || speed < 1 ? 2 : speed
     };
